+++ src/commands/pull.ts
@@ -4,11 +4,8 @@ import { createApiClient } from "../lib/api/client";
 import * as fs from "fs";
 import * as path from "path";
 import shell from "shelljs";
-import {
-  createDecipheriv,
-  CipherKey,
-  BinaryLike,
-} from "crypto";
+import { GithubConnection } from "../lib/api";
+import { decrypt } from "../lib/utils/git";
 
 export default class Pull extends BaseCommand {
   static description =
@@ -191,13 +188,6 @@ export default class Pull extends BaseCommand {
       // Fetch GitHub connection details
       this.log("");
       this.log("Fetching GitHub connection details...");
-      /* eslint-disable camelcase */
-      interface GithubConnection {
-        id: number;
-        accessToken: string;
-        github_username: string;
-      }
-      /* eslint-enable camelcase */
 
       let githubConnection: GithubConnection;
       try {
@@ -215,7 +205,7 @@ export default class Pull extends BaseCommand {
       const branch = link.githubBranch || "main";
 
       // Decrypt token for authentication
-      const decryptedToken = this.decrypt(githubConnection.accessToken);
+      const decryptedToken = decrypt(githubConnection.accessToken);
       const authenticatedRepoUrl = `https://x-access-token:${decryptedToken}@github.com/${link.githubOwner}/${link.repoName}.git`;
 
       this.log("");
@@ -289,40 +279,6 @@ export default class Pull extends BaseCommand {
       this.warn(`Failed to pull service code: ${err.message}`);
       this.warn("You can still use 'apso push' to push schema changes.");
     }
-  }
-
-  /**
-   * Decrypt GitHub access token
-   * TODO: Move this to a shared utility
-   */
-  private decrypt(token: string): string {
-    const aesKeyB64 =
-      process.env.AES_KEY_B64 ||
-      "7N0eyS0YaZXKlXBK+tSY+3i/tKrKWqjrwaK++XtJSn8=";
-    const key = Buffer.from(aesKeyB64, "base64");
-    const [ivB64, dataB64, tagB64] = token.split(".");
-    if (!ivB64 || !dataB64 || !tagB64) throw new Error("Invalid token format");
-
-    const iv = Buffer.from(ivB64, "base64url");
-    const encrypted = Buffer.from(dataB64, "base64url");
-    const tag = Buffer.from(tagB64, "base64url");
-
-    const decipher = createDecipheriv(
-      "aes-256-gcm",
-      key as CipherKey,
-      iv as BinaryLike,
-      {
-        authTagLength: 16,
-      }
-    );
-
-    decipher.setAuthTag(tag as NodeJS.ArrayBufferView);
-
-    const decrypted = Buffer.concat([
-      decipher.update(encrypted as any),
-      decipher.final(),
-    ] as any);
-    return decrypted.toString("utf8");
   }
 }
