+++ src/commands/push.ts
@@ -5,19 +5,15 @@ import {
   readProjectLink,
 } from "../lib/project-link";
 import { createApiClient } from "../lib/api/client";
-
 import { getServiceCodeDir, getAuthoritativeApsorcPath, syncApsorcToCodeBundle, getProjectRoot } from "../lib/project-link";
 import * as fs from "fs";
 import * as path from "path";
 import shell from "shelljs";
-import {
-  createDecipheriv,
-  CipherKey,
-  BinaryLike,
-} from "crypto";
+
 import AdmZip from "adm-zip";
 import FormData from "form-data";
-
+import { GithubConnection } from "../lib/api";
+import { decrypt } from "../lib/utils/git";
 export default class Push extends BaseCommand {
   static description =
     "Push local .apsorc schema to the platform";
@@ -172,13 +168,6 @@ export default class Push extends BaseCommand {
 
       // Fetch GitHub connection details
       this.log("  Fetching GitHub connection details...");
-      /* eslint-disable camelcase */
-      interface GithubConnection {
-        id: number;
-        accessToken: string;
-        github_username: string;
-      }
-      /* eslint-enable camelcase */
 
       let githubConnection: GithubConnection;
       try {
@@ -196,7 +185,7 @@ export default class Push extends BaseCommand {
       const branch = link.githubBranch || "main";
 
       // Decrypt token for authentication
-      const decryptedToken = this.decrypt(githubConnection.accessToken);
+      const decryptedToken = decrypt(githubConnection.accessToken);
       const authenticatedRepoUrl = `https://x-access-token:${decryptedToken}@github.com/${link.githubOwner}/${link.repoName}.git`;
 
       this.log(`  Repository: ${link.githubOwner}/${link.repoName}`);
@@ -301,40 +290,6 @@ export default class Push extends BaseCommand {
       this.warn(`Failed to push service code: ${err.message}`);
       this.warn("Schema push succeeded, but code push failed.");
     }
-  }
-
-  /**
-   * Decrypt GitHub access token
-   * TODO: Move this to a shared utility
-   */
-  private decrypt(token: string): string {
-    const aesKeyB64 =
-      process.env.AES_KEY_B64 ||
-      "7N0eyS0YaZXKlXBK+tSY+3i/tKrKWqjrwaK++XtJSn8=";
-    const key = Buffer.from(aesKeyB64, "base64");
-    const [ivB64, dataB64, tagB64] = token.split(".");
-    if (!ivB64 || !dataB64 || !tagB64) throw new Error("Invalid token format");
-
-    const iv = Buffer.from(ivB64, "base64url");
-    const encrypted = Buffer.from(dataB64, "base64url");
-    const tag = Buffer.from(tagB64, "base64url");
-
-    const decipher = createDecipheriv(
-      "aes-256-gcm",
-      key as CipherKey,
-      iv as BinaryLike,
-      {
-        authTagLength: 16,
-      }
-    );
-
-    decipher.setAuthTag(tag as NodeJS.ArrayBufferView);
-
-    const decrypted = Buffer.concat([
-      decipher.update(encrypted as any),
-      decipher.final(),
-    ] as any);
-    return decrypted.toString("utf8");
   }
 
   private async zipServiceCode(codeDir: string, workspaceId: string, serviceId: string, api: ReturnType<typeof createApiClient>): Promise<void> {
